django-rest-framework/hello_world/requirements.txt
Django==2.1.5
djangorestframework==3.9.1
